DataSHIELD login details builder

DataSHIELD login details builder

DataSHIELD login details builder

Format

A R6 object of class DSLoginBuilder

Details

Helper class for creating a valid data frame that can be used to perform datashield.login. See also newDSLoginBuilder.

Methods

Public methods

DSLoginBuilder$new()

DSLoginBuilder$append()

DSLoginBuilder$build()

DSLoginBuilder$clone()

Method new()

Create a new DSLoginBuilder instance.

Usage

DSLoginBuilder$new(logins = NULL, .silent = FALSE)

Arguments

logins

A valid login details data frame to initiate the builder, optional.

.silent

Do not warn user when non secure HTTP urls are encountered. Default is FALSE.

Returns

A DSLoginBuilder object.

Method append()

Append login information for a specific server.

Usage

DSLoginBuilder$append(
  server,
  url,
  table = "",
  resource = "",
  driver = "OpalDriver",
  user = "",
  password = "",
  token = "",
  options = "",
  profile = ""
)

Arguments

server

The server name (must be unique).

url

The url to connect to the server or a R symbol name.

table

The table path that identifies the dataset in the server.

resource

The resource path that identifies the resource reference in the server.

driver

The DSDriver-class name to build the DSConnection-class.

user

The user name in the user credentials.

password

The user password in the user credentials.

token

The personal access token (ignored when user credentials are not empty).

options

Any options (R code to be parsed) that could be relevant for the DS connection object.

profile

The DataSHIELD R server profile (affects the R packages available and the applied configuration). If not provided or not supported, default profile will be applied.

Method build()

Build the DSLoginBuilder instance.

Usage

DSLoginBuilder$build()

Returns

The DataSHIELD logindata data.frame

Method clone()

The objects of this class are cloneable with this method.

Usage

DSLoginBuilder$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.