Expression result assignment

Assign the result of the execution of an expression to a R symbol in the Datashield R session.

datashield.assign.expr(conns, symbol, expr, async = TRUE)

Arguments

conns
DSConnection-class object or a list of DSConnection-classs.

symbol
Name of the R symbol.

expr
R expression with allowed assign functions calls.

async
Whether the result of the call should be retrieved asynchronously. When TRUE (default) the calls are parallelized over the connections, when the connection supports that feature, with an extra overhead of requests.

Examples

if (FALSE) {
# assign an expression to G
datashield.assign.expr(conns, symbol = "G", expr = quote(as.numeric(D$GENDER)))

# assign the expressions that are defined in the provided named list. 
# Connections are filtered by the list names.
datashield.assign.expr(conns, "G",
  list(server1=quote(as.numeric(D$GENDER)), server2=quote(as.numeric(D$SEX))))
}