Get DataSHIELD-aware data repository metadata

Get DataSHIELD-aware data repository metadata

dsGetInfo(dsObj, ...)

Arguments

dsObj
An object inheriting from DSObject-class, i.e. DSDriver-class, DSConnection-class, or a DSResult-class.

...
Other arguments to methods.

Value

a named list

Implementation notes

For `DSDriver` subclasses, this should include the version of the package (`driver.version`) and the version of the underlying client library (`client.version`).

For `DSConnection` objects this should report the version of the data repository application (`repo.version`) and its name (`repo.name`), the database name (`dbname`), username, (`username`), host (`host`), port (`port`), etc. It MAY also include any other arguments related to the connection (e.g., thread id, socket or TCP connection type). It MUST NOT include the password.

For `DSResult` objects, this should include the R expression being executed (an expression object tailored by the implementation of DSI) and if the query is complete (a result object tailored by the implementation of DSI).

See also